Crate pallet_protos

This pallet protos performs logic related to Proto-Fragments.

IMPORTANT STUFF TO KNOW:

A Proto-Fragment is a digital asset that can be used to build a game or application

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

AccountsInfo

Struct that represents the account information of a Proto-Fragment

GetProtosParams

Data Type used to Query and Filter for Proto-Fragments

Proto

Struct of a Proto-Fragment

ProtoPatch

Struct of a Proto-Fragment Patch

Enums

LinkSource

TODO: Documentation

LinkedAsset

Types of Assets that are linked to a Proto-Fragment (e.g an ERC-721 Contract etc.)

ProtoOwner

Types of Proto-Fragment Owners

UsageLicense

Enum that indicates how a Proto-Fragment can be used

Traits

WeightInfo

Weight functions needed for pallet_protos.