Module pallet_protos::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configure the pallet by specifying the parameters and types on which it depends.

Type Definitions

MetaKeys

StorageMap that maps a Metadata Key (of type Vec<u8>) to an index number

MetaKeysIndex

StorageValue that equals the total number of unique Metadata Keys in the blockchain

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

Protos

StorageMap that maps a Proto-Fragment’s data’s hash to a Proto struct (of the aforementioned Proto-Fragment)

ProtosByCategory

StorageMap that maps a variant of the Category enum to a list of Proto-Fragment hashes (that have the aforementioned variant)

ProtosByOwner

StorageMap that maps a variant of the ProtoOwner enum to a list of Proto-Fragment hashes (that have the aforementioned variant)

ProtosByParent

StorageMap that maps a Proto-Fragment to a list of other Proto-Fragments that reference the Proto-Fragment

Tags

StorageMap that maps a Tag (of type Vec<u8>) to an index number

TagsIndex

StorageValue that equals the total number of unique tags in the blockchain

TraitImplsByShard

StorageMap that maps a Shards-Proto (i.e a Proto-Fragment that is a Shards Script) to all the traits that are in the “implementing” field (ShardsScriptInfo::implementing) of the Shards-Proto’s category or in the “implementing” field of one Shard-Proto’s Shard-Proto ancestor’s category

Traits

StorageMap that maps a Trait ID to the name of the Trait