Trait pallet_protos::pallet::Config

pub trait Config: Config + Config + Config + Config + Config + Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type WeightInfo: WeightInfo;
    type StringLimit: Get<u32>;
    type DetachAccountLimit: Get<u32>;
    type MaxTags: Get<u32>;
}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type WeightInfo: WeightInfo

Weight functions needed for pallet_protos.

type StringLimit: Get<u32>

The maximum length of a metadata key or a proto-fragment’s tag or a fragment definition’s name that is stored on-chain.

type DetachAccountLimit: Get<u32>

The maximum length of an Public Account Address on an External Blockchain that can be given sole ownership of a Proto-Fragment or a Fragment Instance.

type MaxTags: Get<u32>

The maximum number of tags that a single Proto-Fragment can be tagged with.