Crate pallet_protos

This pallet fragments performs logic related to Proto-Fragments.

IMPORTANT STUFF TO KNOW:

A Proto-Fragment is a digital asset that can be used to build a game or application

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

AccountsInfo

GetProtosParams

Data Type used to Query and Filter for Proto-Fragments

Proto

Struct of a Proto-Fragment

ProtoPatch

Struct of a Proto-Fragment Patch

Enums

LinkSource

¿

LinkedAsset

Types of Assets that are linked to a Proto-Fragment (e.g an ERC-721 Contract etc.)

ProtoOwner

Types of Proto-Fragment Owners

Traits

WeightInfo

Weight functions needed for pallet_protos.