Trait pallet_protos::pallet::Config

pub trait Config: Config + Config + Config + Config + Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type WeightInfo: WeightInfo;
    type StorageBytesMultiplier: Get<u64>;
    type CurationExpiration: Get<u64>;
    type TicketsAssetId: Get<<Self as Config>::AssetId>;
}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type WeightInfo: WeightInfo

Weight functions needed for pallet_protos.

type StorageBytesMultiplier: Get<u64>

Weight for adding a a byte worth of storage in certain extrinsics such as upload().

type CurationExpiration: Get<u64>

The number of blocks after which a curation period is over

type TicketsAssetId: Get<<Self as Config>::AssetId>

Asset ID of the fungible asset “TICKET”